Frame Analysis: Make deffered_ctx_accurate work with stereo
The driver's reverse stereo blit control does not apply to a specific
context - rather, it is per device/stereo handle, and enabling it while
queuing commands for a reverse stereo blit operation in a deferred
context doesn't work.

However, if the option is enabled when the deferred context's command
list is executed in the immediate context, then any reverse stereo blit
operations do work.

This also disables any attempt to set it when using a deferred context
to dump resources (except when deferred_ctx_immediate is enabled) so
that they cannot interfere with the reverse stereo blit control on the
immediate context.