Frame Analysis: Fix broken Texture2D hash causing many collisions
When examining a frame analysis dump of a full frame, I noticed that
many of the images that should have had an update were not showing it,
and this seemed to correspond to changed in the image that should have
occurred in the mid to lower areas of the image.

We've known for quite a while now that our texture hash was broken, and
did not consider the entire image (depending on the format), but this is
ridiculously bad.

Add a variant of CalcTexture2DDataHash that tries to calculate the hash
accurately - using the full image, and skipping over any padding inside
the image. Hopefully (assuming there are no other bugs) this will be a
canonical texture hashing implementation we can use, and perhaps
consider switching the main texture hashes to it in the future.