Resource Hashes: Add option to use accurate Texture2D hashes
This adds an option to change the texture hashes to use the new more
accurate version. This isn't enabled by default because it would break
too many fixes, but comparing the results with frame analysis
deduplication shows some pretty compelling reasons to consider using
this:

- The old hashes wouldn't hash the entire image, so if two textures only
  began to differ half way down the image, their hashes would collide.

- The old hashes were susceptible to garbage in the padding between rows
  that could alter the hashes.

- The new hashes mean the filenames in the frame analysis deduped folder
  will match the texture hashes for static textures dumped as mono.
  Prior to this change, less than half of the hashes we expected to
  match in DOAXVV actually did, after this change they all do (hash
  tracking is also required in that game). Render targets are not
  expected to match, since they will have legitimately been altered.

- The new hash format could potentially be calculated on demand,
  potentially giving us an alternative to hash tracking in the future,
  though the performance of that is still uncertain.

Leave the option disabled by default for compatibility with existing
fixes, and because the new scheme will hash the entire texture, which
will naturally take longer than the old scheme of not hashing the entire
texture.