Fix regression preventing full screen override in Batman: Telltale
Batman: Telltale passes in NULL for pFullscreenDesc to
CreateSwapChainForHwnd, and because we were trying to avoid possible
behaviour changes of the original call (like not returning
DXGI_ERROR_INVALID_CALL if the original call was expected to) by
changing NULL parameters to non-NULL, we would skip the full screen
override code.

Since passing NULL for pFullscreenDesc should be equivalent to passing a
filled in structure with Windowed = true, we should be able to always
create our own filled in structure and pass that to the original call
without changing its behaviour compared to passing in NULL, and this
allows to keep the code relatively simple, but allowing the full screen
override to work regardless of whether the game passed in a structure or
NULL for pFullscreenDesc.

Reported-by: Bo3b