Command List: Add "stereo_active" ini param override
This fits in with the other ini param overrides that get the current
stereo state. This one in particular is to be used by the
auto-convergence shader to detect if stereo was active in the previous
frame to prevent interpreting alt+tab as the user adjusting the
convergence. When stereo is inactive, the auto convergence shader won't
be run in the left eye and cannot update the left eye's copy of its
state buffer, and this gives it a facility it can use to detect this.