Frame Analysis: Use the view format to avoid dumping TYPELESS resources
Games tend to use TYPELESS formats for resources, as this gives them
more flexibility in how they can use the resources (e.g. this is
necessary to use a depth buffer in a shader resource slot), and instead
they assign a fully typed format to each view of the resource.

Frame analysis uses the format of the resource when dumping files, but
this means that the resulting DDS files will often be TYPELESS, and
external programs (including Microsoft's texconv from DirectXTex) won't
know how to interpret them, or will have to guess at what the format
really is.

To improve compatibility of the dumped DDS files with third party
programs, this sets the format of the staging resource to that obtained
from the view wherever possible. Since this changes the binary contents
of the file but doesn't change the resource hash (which does use the
resource format, but not the view format), the format is appended to the
deduped filename.

As an added bonus, this means that DirectXTK will now be able to dump
many more resources as JPEG, providing another massive speed boost and
disk space savings.