Fix Overlay and Swap chain preventing each other from being released
The overlay holds a reference on the swap chain, which will prevent the
swap chain from being released until the overlay has been deleted, but
the overlay will not be deleted until the swap chain is released, so
having the overlay hold a reference on the swap chain prevents
everything from being released.

If the overlay still exists we know the swap chain hasn't been released
yet, so there is no need to hold a reference on it.