Ini Handler: Fix refcounting bug and incorrect context
The Ini Handler calls GetImmediateContext to get the immediate context
to update iniParams, but failed to balance this call with a release,
preventing the context from being released if the config had ever been
reloaded.

It also called Map/Unmap on the HackerContext returned from
GetImmediateContext, rather than bypassing it and calling through to
DirectX, which could potentially allow the HackerContext's map/unmap
code to interfere with updating iniParams.