Clear HackerDevice's reference to HackerSwapChain on Release
The device doesn't hold strong references to children objects such as
the immediate context and swap chain (which is correct, as doing so
would prevent anything from being released), but it does sometimes need
to access them, so it still holds a pointer to them (which effectively
amounts to a weak reference).

But if the child object is released the device should cease accessing
these child objects as the memory they point to is no longer valid. We
already handle this for the immediate context, but were not doing this
for the swap chain.

This NULLs out the HackerDevice's pointer to the HackerSwapChain when
the later is released, and adds NULL pointer checks to any code that
accesses these.