HackerSwapChain: Hold refs to the device & context (+ fix refcounting bug)
Since the HackerSwapChain uses the HackerDevice and HackerContext really
it should own the references to them (either instead of or as well as
the overlay) to make sure they can't be released prematurely.

This fixes a refcounting bug, where if the swap chain was created with
no immediate context it would use GetImmediateContext to create, wrap
and hold a reference to the immediate context, that it never released -
this reference will now correctly be released when the swap chain is
released.