Overlay: Be consistent about which interface we are AddRef()ing & Release()ing
The overlay calls AddRef on the HackerInterfaces, but Releases the
DirectX interfaces. These should be consistent, and we should be
performing this on the Hacker interfaces since bypassing the Hacker
interfaces for the Release call could result in a leak and a stale
HackerContext still accessible via GetImmediateContext (there are likely
other possible circumstances where a similar issue could arise).