Make [Constants] a command list for one time initialisation
The [Constants] section defines the state of IniParams on initial load
and config reload, but there may be cases where we want to initialise
additional state as well. To facilitate this, convert [Constants] into a
command list that is run when the immediate context is created and
whenever the config is reloaded. Since the command list supports a
superset of the syntax used in the [Constants] section, this should be
backwards compatible with existing fixes, but gives us maximum
flexibility in what can be initialised from this section.