Overlay: Back up and restore pixel shader constant buffer 0
The dark background of the notifications uses DirectXTK to draw a
rectangle, which as it turns out will do a PSSetConstantBuffers() call
that was not done when we were only using it to draw text.

Back up and restore the affected constant buffer to fix rendering issues
in certain games, such as Elite Dangerous if the overlay is displayed on
launch.