Command List: Have stereo overrides not require "pre" in [Constants]
The separation & convergence overrides ordinarily require "pre" or
"post" specified to make them apply to more than just the current draw
call, but since we aren't adding a "post" command list for [Constants]
this is not possible. It doesn't make sense to restore them on post in
this case anyway, so always disable restore_on_post in the event that no
post command list exists.