Frame Analysis: Fix edge case where share_dupes wouldn't work on a deferred context
The currently active analyse_options is normally stored in the
FrameAnalysisContext and updated throughout the frame as triggers affect
it or for the per-frame dump command, but when dumping resources that
were delayed due to being used in a deferred context we need to use the
analyse_options that was saved along with the resource, and to this end
we had a local copy of analyse_options passed into the immediate context
dumping functions, but this local copy of analyse_options was not used
by any of the helper functions they could call - the dedupe_filename
routines in particular.

This could lead to an edge case where share_dupes would be ignored if it
was not set in the global analyse_options, but was set in a per-shader
analyse_options or dump.

To keep things simple and maintain the original semantics, instead of
passing a local copy of analyse_options into the immediate context
dumping functions, set it in the FrameAnalysisContext before calling
them, so that they and any helpers they use will use the correct value.