Command List: Have dump command pass format to Frame Analysis to avoid TYPELESS dumps
The dump command already passes a format to frame analysis, but the only
cases where it is filled out is for index buffers and custom resources.
In all other cases, FillInMissingInfo() is required to fill it in from
the view.

By adding this, frame analysis dumps triggered from the command list are
now dumped as fully typed, giving the same improved third compatibility
as outlined earlier, and increasing the number of resources we can jump
as jps.

FillInMissingInfo is moved up in the file to avoid the need for a
prototype, but is unchanged.