Frame Analysis: Allow user to abort frame analysis & report completion
Since Frame Analysis can take a long time depending on what options are
enabled, it can be pretty costly to activate by mistake, especially when
dumping from deferred contexts where running out of memory is a real
possibility.

This gives the user an opportunity to abort a frame analysis session in
progress - it calls into the input subsystem to process key inputs
before dumping a resource and will abort if frame analysis has been
disabled, which can now be done by pressing the activation key a second
time (or releasing it if the "hold" frame analysis option is enabled).

Frame analysis will now report why it completed to the overlay, so the
user can see at a glance whether it finished a frame, was aborted by
pressing the button again, or how many frames they caught with hold
mode.