Overlay: Add an outline around the main overlay text
In most games the main overlay text is fairly readable already, but
occasionally a game might use bright pastel colours making it harder to
read. It probably doesn't need the dark rectangle background of the
notifications, but outlining the letters can help.

DirectXTK doesn't have a built in means of doing this, but their
documentation recommends drawing the text multiple times at different
offsets, so that's what we do.