Overlay: Increase convergence precision to 2 decimal places
Convergence changes of less than 0.1 can be pretty significant, and does
have a noticeable effect on the stereo image. Since people tend to use
the overlay to get the convergence for the profile and presets, increase
its precision to 2 to help lock it in more precisely. There's not much
point going beyond this, as a convergence change of less than 0.01 is
barely noticeable.