Command List: Fix draw=from_caller when drawing nothing at all
This is seen in Senran Kagura Peach Beach Splash:

	000219 DrawIndexed(IndexCount:0, StartIndexLocation:0, BaseVertexLocation:0)
	000219 3DMigoto pre {
	000219 3DMigoto  [ShaderOverrideHUDText] run = customshaderhudtext
	000219 3DMigoto  pre {
	000219 3DMigoto   [customshaderhudtext] Draw = from_caller -> DrawAuto()
	000219 3DMigoto   [customshaderhudtext] handling = skip
	000219 3DMigoto  }
	000219 3DMigoto }

If the game tries to draw nothing and we redirect it to a custom shader
via draw = from_caller, the command list will observe vertices=0,
indices=0 and instances=0 and conclude that the only draw call that
makes sense in that case is DrawAuto(), which while true, is not what
the game did.

We aren't going to draw anything in this case, and chances are that
DrawAuto will not do anything either since no stream output has been
bound back to an input, but we shouldn't call it in case it has some
unexpected side effect.

Change the draw context structure to explicitly include which draw call
was made, and have the command list use that instead of implying the
draw call from which arguments are in use.