Frame Analysis: Make symbolic links optional
It seems that despite the fact that Microsoft has added support for
symbolic links, they are still stuck in a mindset of thinking they need
to handle them specially as they would for windows shortcuts, but doing
so goes against the design principles of symbolic links - the only
reason they should use special code for symbolic links is to manipulate
the link itself, and in all other cases they should treat it as a
regular file.

The result of this is that windows explorer fails to render thumbnails,
and Windows photo viewer can open the files, but cannot navigate between
them. These are both examples of bugs resulting from trying to treat the
symbolic link specially and making a mess of it - if these applications
were not symlink aware these cases would both just work.

Rant over. Make symbolic links optional so that hard links will be used
by default, since just disabling developer mode won't prevent 3DMigoto
from using them if the game is running as admin.

Reported-by: DJ-RK