Shaderusage: Fix possible crash if D3D11_KEEP_UNORDERED_ACCESS_VIEWS
Our OMSetRenderTargetsAndUnorderedAccessViews code would update our
local copy of the number of UAVs and UAV start slot, even when NumUAVs
was passed as D3D11_KEEP_UNORDERED_ACCESS_VIEWS (0xffffffff). This would
then be used by RecordGraphicsShaderStats as though it was the number of
UAVs, leading to a bad DirectX call, buffer overrun and possible crash.

This makes the stat collection in this call check for the keep options
of RTVs/DSV and UAVs separately, and only updates the stats if they are
not set to keep the existing views.

This may fix the crash in Shadow of War:
https://github.com/bo3b/3Dmigoto/issues/91