Blacklist QueryInterface(IDXGISwapChain4)
Shadows of War is seen requesting this on Windows 10 shortly before a
crash. Since we don't yet support this interface we need to blacklist
it, however this just results in another error in Shadows of War (this
time an actual error message about not being able to create an
IDXGISwapChain rather than a crash).

Blacklisting is still the right thing to do for now - once we switch to
the vs2015 branch we can look at adding actual support for higher level
DirectX objects including this one.