Frame Analysis: Leave one free hard link on each file
Moving files from one directory to another in cygwin throws "too many
hard links" errors when moving multiple files that have the maximum
number of hard links from one directory to another (the first file is
moved successfully, additional files sharing the same inode are
attempted to be temporarily hard linked to the first before the original
is removed, which fails when the files already have the maximum number
of hard links). Retrying the move many times will eventually succeed,
but this only allows one extra file moved on each attempt.

Although this is pretty clearly a cygwin issue, we can play nice and try
to prevent generating files that are already at their hard link limit,
by checking how many links they have, and rotating them when they are
nearing the limit. This rotates files once they reach 1023 links
(including the master), one below the NTFS limit of 1024. cygwin can
move these files without issues.

The 1023 limit for NTFS is currently hard coded, but since that is a
filesystem limit and may not be constant (e.g. it could potentially be
lower on a Samba share, or if a third party filesystem driver is in
use), the old rotate on failure code path is also retained.