Presets: Delay activation until present call
The presets are currently activated the moment they are triggered, which
is generally fine as the actual changes are scheduled to occur later on
the following Present call(s), but since we can now trigger a command
list from these, that could occur during the frame.

In most cases this won't matter too much, but that command list really
should be triggered when the preset takes effect, which will be during
the present call, and not during the middle of the frame. Since the
preset is itself triggered from a command list, any commands that are
supposed to run in the middle of the frame can just be run from there,
so this doesn't take away anything.