Overrides: Do not run post command list if not active & refactor to simplify
If an override is not activated due to its condition not being met, we
don't want to run the post command list when it is deactivated.

Refactor the active/inactivate tracking into the Activate/Deactivate
functions themselves, and out of the Toggle() function and the
PresetOverride class, allowing this to be tested regardless of the
override type (provided said type has a deactivate condition).

Refactor the global save/restore into the Activate/Deactivate functions
- Deactivate always restores (decreases the refcount of) the global
state provided the override was active, and Activate only saves the
global state if the override type has a corresponding deactivate
condition.

Refactors the now one-liner PresetOverride::Activate/Deactivate
functions back into PresetOverride::Update.