Input: Allow multiple keys for each [Key] section
This allows a single [Key] section to be activated by multiple key
bindings, which is useful for toggle or cycle type bindings so that the
current state can be shared regardless of key.

e.g. Instead of writing two key bindings for keyboard and xbox
controller:

    [KeyKB]
    key = x
    x = 1, 2, 3, 4, 5

    [KeyXB]
    key = XB_LEFT_THUMB
    x = 1, 2, 3, 4, 5

They can now be combined into a single section:

    [KeyKBandXB]
    key = x
    key = XB_LEFT_THUMB
    x = 1, 2, 3, 4, 5

And the state of the cycle will be shared between both keys.