Input: Add wrap option to type=cycle keys
Now that we can traverse forwards or backwards on type=cycle key
overrides, add a wrap option to indicate what happens when the cycle
reaches either end of the list. By default it will wrap from the end
back to the start when cycling forwards, or from the start back to the
end when cycling backwards. Setting wrap=false prevents this, and the
key binding will have no effect if trying to cycle off either end of the
list.