Update dxgi loader for Win10 support. Changes the .def file to avoid the unnecessary @x imports, and just use named functions.  Add the needed CreateDXGIFactory2 export for the dll.  The other Win10 specific functions are documented, but not added, because we have no use case.
Add the CreateDXGIFactory2 function to the cpp and implement it the
same as the prior two, noting that it has a different signature.

Added a static ID3D11CreateDevice reference to the project file, so
that any load of this dxgi.dll will also statically load d3d11.dll
automatically.  Removed calling LoadLibrary, as this is more reliable.

Tested in ShadowOfWar on Win10, and successfully loads our d3d11.dll.