Command List: Add direct_mode_eye command
This command will set the active eye in 3D Vision Direct mode. It can be
used to render 3DMigoto overlays in both eyes, copy the back buffer from
the previous frame to the next (for sgsrules Unity plugin), etc.

There is currently no way to retrieve the currently active eye from the
command list (e.g. to restore the original setting once we are done), on
account of nvapi not providing an API to do this. We could potentially
use our nvapi wrapper to get this as the game sets it, or hook into the
call from our d3d11 wrapper. For now this doesn't seem to be required
for sgsrules tool.