Frame Analysis: Dump vertex buffer layouts
This dumps the contents of the input assembler layout along with the
vertex buffers to assist in decoding them. This currently just dumps the
layout as text, and does not use it to help decipher the vertex buffers.

DirectX has no API to retrieve the layout description from the layout
objects, so we create a copy of the layout description at the time of
creation and associate it with the layout object's private data. We use
an ID3D11Blob to wrap this, as that takes care of releasing the memory
when the layout interface is released. The private data is known to be
slow to access, but this doesn't matter since this is only used for
frame analysis where performance isn't critical.