Frame Analysis: Determine maximum vertex count from index buffer
When we dump vertex buffers as text from an indexed draw call we only
know the vertex buffer offset & base vertex index, but not the maximum
vertex index, and as a result we would dump the entire vertex buffer
from the starting position onwards. This worked, but in some games that
combined many objects into a single vertex buffer (such as Witcher 3)
would result in many copies of a single vertex buffer only varying by
starting offset, and with a lot of irrelevant vertices after the ones we
are interested, that in some cases could even be complete garbage as
they followed a different layout. These files could easily be 10MB or
more, often making them the largest disk space consumer in the dump and
the deduping process could only reduce this so far.

To address this, we scan over the index buffer used in the draw call to
find the largest vertex index and use that to determine the number of
vertices we will actually need to dump.

Since the index buffer isn't directly available to the CPU, we need to
read it after it has been staged. To this end we make "dump_vb txt"
imply "dump_ib" so that we will stage it and have DumpBuffer() be able
to pass a reference to the staged index buffer back to the caller. This
staged index buffer (along with the rest of the details necessary to
decode it) is then passed into the DumpBuffer() routine when dumping the
vertex buffers and is used to determine the vertex count.