Frame Analysis: Fix dumping of per-instance data in vertex buffers
Vertex buffers also double as instance buffers, which can be at a
different offset in the same buffer, or in a separate buffer. Either way
the per-instance data has separate first and count entries to the
per-vertex buffer, and some other quirks, like being able to repeat
individual per-instance elements over several instances before moving on
to the next element in the buffer.

To this end, if we are dumping a vertex buffer and know the layout we
now handle the per-instance data separately from the per-vertex data
using the correct first and count parameters from the draw call, and use
the InstanceDataStepRate so that each element of per-instance data can
be advanced separately.