Frame Analysis: Dump primitive topology
This outputs the primitive topology in the vertex & index buffers, as
this is one of the few remaining pieces necessary to decode these
standalone, e.g. for import into an external tool.

It also makes the index buffer dumps a little nicer for triangle & line
lists, dumping the indices in groups of 3 or 2 on each line respectively.