Command List: Allow hunting skip to affect all custom draw commands
We already have provisions to skip custom draw commands in command lists
if the object that triggered the command list was skipped due to
hunting, but that only applied to draw=from_caller. However, in some
cases we might use one of the other variants of the custom draw call
(e.g. to override the number of indices for mesh replacement) and in
that case we still want hunting to work, so make all custom draw calls
affected by the hunting skip.