Set rasterizer_disable_scissor=0 by default
While we usually want to disable scissor clipping in stereo, doing so
universally can cause other issues, and since these do not show up with
the show_original key, it is not even always obvious that anything has
gone wrong or what exactly is at fault.

This has come up several times lately:
- Breaks shadow maps in Dead or Alive Xtreme Venus Vacation (allows
  shadows to extend into unrelated cascaded shadow maps, resulting in
  strange lines on the ground)
- Breaks minimap in .hack//G.U. https://github.com/bo3b/3Dmigoto/issues/90
- https://forums.geforce.com/default/topic/685657/3d-vision/3dmigoto-now-open-source-/post/5315705/#5315705

In general we don't want 3DMigoto to alter a game's appearance until a
modder intentionally does so, and having scissor clipping disabled by
default is counter to that goal.

Another point to consider is that rasterizer_disable_scissor requires a
game restart to reliably take effect, so it's a hassle if we find we
need to change it. However, 3DMigoto nowadays has "disable_scissor=1",
which can be applied on a per-shader basis and does *not* require a game
restart to take effect, and will respect the show_original key.

The downside of this is that scissor clipping quite often *is* broken in
3D and shaderhackers will have one more thing to look out for and know
how to correct - but is that actually any better than the non-obvious
ways disabling scissor clipping can break effects?

So, even though going back to enabling scissor clipping by default means
some more work for the shaderhackers, it is in many ways superior to
disabling scissor clipping by default and having 3DMigoto potentially
break a game just by "being installed".