Fuzzy TextureOverrides: Add draw call context matching
This is a little different to the other fuzzy matches, as it is not
checking any part of the resource description (and therefore can be used
with either hash or fuzzy resource description matching), but rather the
context of the active draw call when invoked from the command list.

It can check the first vertex or index to match a sub-part of a mesh, or
a specific mesh inside a larger common buffer. It can check the first
instance to match a specific set of instances of a given mesh, and can
check the vertex/index/instance count to match specific types of draw
calls or otherwise limit what meshes it can match.