TextureOverride Draw Call Context Matching: Allow Duplicate Hashes
If we are matching the sub-parts of a mesh, the hash can no longer be
treated as a unique key for the related TextureOverride sections.

We don't want to totally remove the meaning of the hash though -
matching on a hash still should preclude all fuzzy matches on resource
descriptions, only now we can potentially have several TextureOverride
sections that also match the same hash, and within that group the same
priority as the fuzzy sections will still need to be applied.

This turns the TextureOverride map into a map of sorted vectors - we
can't use sorted sets, because sets have to be const and we have code
that mutates the TextureOverride structures after creation (not in any
way that would change the custom sorting mind you). Instead we sort the
vectors as they are created in the ini handler.