Command List: Add drawindexed=auto to draw all indices in the current buffer
This command is a convenience for when every index in the current index
buffer needs to be drawn, without having to know in advance how many
that is. It is only suitable if there is only a single object in the
index buffer starting at the first index and filling the entire buffer,
but if we are loading custom index buffers than this will generally be
the case (and if it's not the case then there is no real way around
specifying the from/count fields manually).