Presets: Add unique_triggers_required option
This option indicates that a preset must be triggered from at least this
many unique command list commands each frame to activate, and can be
used to narrow down the circumstances in which a preset might activate
by requiring multiple conditions.

The reason for it requiring multiple *unique* commands is to combat
cases where an effect used for the trigger may use many and variable
draw calls, which cannot be predicted in advance.