Custom Resources: Allow textures loaded from file to override bind_flags
When a custom resource is loaded from a file, the bind_flags override is
ignored, which limits us in certain special circumstances, such as if
that resource is only copied by reference into another resource and
never directly bound to the pipeline itself, or when we want to
manipulate driver heuristics.

We can't override the full resource description for textures loaded from
disk, since DirectXTK doesn't give us a chance to, but it does allow us
to pass in the bind flags we want to use, so allow them to be
overridden.