Frame Analysis: Make the dump command imply deferred_ctx_accurate
If the dump command is used the user probably expects it to just work,
but if the resource is in a deferred context it won't unless they
specify one of the deferred context dumping options as well, and
that can be easy to forget and leads to a moment of confusion.

We can't enable deferred context dumping by default globally because it
would run out of memory and crash, but since the dump command is only
used in limited circumstances it should be relatively safe to enable by
default for that command.