Allow additional configuration files to be included
This adds support for including additional configuration files - both
including specific files by name and recursively including all ini files
in an entire directory is supported.

There are several use cases for this - for starters it can be used to
relocate the configuration of the standard shaders we ship into their
own configuration file, cleaning up the d3dx.ini and allowing them to be
enabled/disabled just by uncommenting their include directive at the top
of the d3dx.ini. The same thing could simplify including the Unity
template, custom debug shaders, etc.

The second and more compelling use case for this support is to allow
3DMigoto to be used as a full modding platform, as opposed to just a
tool for single stand-alone 3D Vision mods. In many aspects 3DMigoto
well and truly surpasses many of the other DX11 graphics modding tools
available, and it makes sense to start pushing it as such. For example,
3DMigoto is well suited to mesh and texture replacements, which is of
interest to certain modding communities involved in things such as
character or costume modding, where 3DMigoto can give them a head start
as the alternative tends to be relying on custom game-specific modding
tools, which often don't exist and would have to be written following
some serious reverse engineering sessions.

A large part of this patch involves adding namespacing support to the
ini sections, so that (non-global) sections included from a separate
configuration file are given a unique name to avoid clashing with
sections in the main d3dx.ini or another mod. In cases where one section
can refer to another (e.g. run=CustomShader/CommandList, preset=,
ShaderRegex*.Pattern, arbitrary resource copying), the namespace is
usually implied by merit of being in the same config file. If the
referenced section is not found the global d3dx.ini namespace will be
checked, and the full namespace can be specified to refer to sections in
a different config file (though this is done by absolute path for now).

This also allows files referenced from these namespaced sections (e.g.
in CustomShader or Resource sections) to be located relative to the
included config file instead of the main 3DMigoto dll, so that these can
be shipped with the mod without the user having to worry too much about
what needs to go where.

Included config files are permitted to append to global sections (such
as [Present]) and override key/value settings in those settings. This
allows a mod that adds an overlay to be largely standalone. Where
ordering between mods is significant (which hopefully most modders will
avoid), 3DMigoto tries to be consistent by sorting the files &
directories case-insensitive alphabetically, to at least be able to give
consistent results.

There are some limitations to this namespacing:

- ShaderFixes is global - if a mod needs files to be placed in here they
  will just have to document as such, and if two mods modify the same
  shader they will conflict. ShaderRegex is a possible alternative, and
  in some cases constant buffer modification can be used instead.

- IniParams is not namespaced, which could easily become a point of
  contention. Solving this with IniParams is non-trivial, but we could
  extend the Custom Resource feature to allow fields inside those to be
  set as easily as IniParams can be.

- ShaderOverride hashes have to be unique, or all sections must indicate
  allow_duplicate_hash=true. I am considering relaxing this restriction,
  or just issuing a short warning and continuing anyway.

- TextureOverride hashes have to be unique (and these lack an
  allow_duplicate_hash setting). I am considering relaxing this
  restriction, but it is also useful to identify conflicts between mods
  (e.g. two mods that replace the same texture or mesh), however even
  then it is not a sure thing since e.g. a mesh may be replaced by
  matching either the index or vertex buffer hash (or potentially even a
  texture the mesh is used with), and those don't conflict with each
  other.