Add per-line namespacing for global sections
Global sections such as [Present] may be appended to from multiple
configuration files with different namespaces, but currently the data
structures only store the ini namespace on a per section basis.

For these sections we need to store the namespace that each line came
from, so that the command list can properly locate sections they refer
to.