Split ini configuration for SBS, Software mouse and upscaling out
This uses the new include functionality to split out the boilerplate for
these three custom shaders into their own configuration files, and
rather than uncommenting a line in the [Present] section to enable them,
a line in the [Include] section at the top of the file is uncommented
instead.

For now I've left the extra configuration that the upscaling shader
needs in the d3dx.ini (that is, upscaling=1, upscale_mode & resolution).
These could be moved out as well, but people and/or 3D Fix Manager may
be looking for them in the d3dx.ini, and setting a default in the
external configuration file would override any in the d3dx.ini

Conversely, hide_cursor is set by the external mouse config file, since
that is always wanted when the software mouse is in use.

The SBS shader has the F11 key moved out, but the x7 definition in the
[Constants] section is left where it is. Again, this is because people /
3D Fix Manager might be looking for it there, and also because it is
important to document that x7 is reserved in the [Constants] section.

Of note - 3D Fix Manager could take advantage of this new functionality
if it wanted to write a user_settings.ini file and have that included
last to override all other settings.