Add option to ignore certain patterns when recursively including files
When shipping mods, it can sometimes be useful to ship an optional part
of a mod that is not enabled by default, but that a user can choose to
enable, or perhaps several choices of a particular type of mod.

One option is just to ship these options separately, but to make things
a little easier this adds an option that can be used to ignore any files
matching a given pattern, e.g. ignore=DISABLED* would ignore any config
files or directories starting with DISABLED, allowing the user to just
rename them to enable that part of the mod.