TextureOverrides: Make match_priority suppress duplicate hash warning
If 3DMigoto is used as a modding platform, there is a relatively high
chance of multiple mods acting on the same texture hash. In some cases
this won't be intentional and would indicate a mod conflict, where the
duplicate hash warning 3DMigoto issues may be the only indication of
what has happened, but in other cases it may be intentional if the mods
do not actually conflict, or if one is intended to override the other.

The existence of match_priority in a TextureOverride signifies some form
of ordering has been defined between sections with the same hash, and
therefore we can use this as a signal to indicate that duplicate hashes
are allowed, even if neither form of fuzzy matching is in use.

Conveniently enough, the recent addition of the fuzzy draw context
matching means that the match_priority field already works to order
TextureOverride sections with the same hash, so all this commit does is
tweak the conditions that we can issue a warning, and to make the
warning clearer if it does appear.