ShaderOverrides: Add allow_duplicate_hash=overrule
allow_duplicate_hash normally has to be set in every ShaderOverride
section that is potentially going to have a duplicate hash to catch
cases where this was by mistake and acknowledge the potential pitfalls
of splitting a section up. It provides allowance for tool-generated
sections to co-exist with manually created sections with the same hash,
while still providing protection against unintended duplicate sections.

However, this doesn't work so well for third party mods, since it
assumes that the mod author will be able to edit mods that are not their
own. There is a legitimate case here for a mod author to just tell
3DMigoto that it's fine.

To facilitate this, we add another possible value to this setting -
"allow_duplicate_hash=overrule", which will stop 3DMigoto from warning
about the duplicate hash regardless of what settings the other sections
with the same hash used (or at least the one parsed just before it and
any parsed thereafter). This isn't a perfect solution, but a perfect
solution is not really viable and this is a reasonable compromise.